deleted first draft to reupload final draft
did not clone this repository and i'm manually saving my jupyter notebooks to upload to github (sorry.. im a little burnt out this week and didn't want to troubleshoot git pushing issues). this final draft does not include the date for the largest water difference in question 3, because i couldnt figure out the equivalent to grep in a python script. i think i have the logic down for how to get the date for this value, but am going to bring this to the help session tomorrow october 20th. 2022